+++ board/src/lib.rs
@@ -85,6 +85,8 @@ pub struct Common {
     pub usbnc1: UsbNc1,
     /// USBPHY1 registers.
     pub usbphy1: UsbPhy1,
+    /// Fuse access through the on-chip one-time programmable controller (OCOTP).
+    pub ocotp: hal::ocotp::Ocotp,
 }
 
 impl Common {
@@ -112,6 +114,8 @@ impl Common {
             ..
         } = hal::snvs::new(unsafe { ral::snvs::SNVS::instance() });
 
+        let ocotp = hal::ocotp::Ocotp::new(unsafe { ral::ocotp::OCOTP::instance() });
+
         Self {
             pit,
             gpt1,
@@ -122,6 +126,7 @@ impl Common {
             usb1: unsafe { Usb1::instance() },
             usbnc1: unsafe { UsbNc1::instance() },
             usbphy1: unsafe { UsbPhy1::instance() },
+            ocotp,
         }
     }
 }