+++ src/chip/drivers/ccm_10xx/clock_gate.rs
@@ -389,6 +389,19 @@ where
     }]
 }
 
+/// Returns the FlexSPI clock gate locator.
+#[inline(always)]
+pub const fn flexspi<const N: u8>() -> Locator
+where
+    ral::flexspi::Instance<N>: ral::Valid,
+{
+    [locator(CCGR6, CG5), locator(CCGR7, CG1)][if N == ral::SOLE_INSTANCE {
+        N as usize
+    } else {
+        N as usize - 1
+    }]
+}
+
 /// Returns the SAI clock gate locators.
 #[inline(always)]
 pub const fn sai<const N: u8>() -> Locator