+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Add OCOTP driver for reading and writing fuses.
 
+Add FlexSPI clock gate locator(s) for 10xx MCUs.
+
 ## [0.5.12] 2025-12-10
 
 Add uSDHC clock gates for 10xx and 11xx MCUs.